backend/requirements.txt
Django==5.1.3
psycopg2==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
qrcode==8.0
Pillow==11.0.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
